% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEzZarrEnvPath.R
\name{getEzZarrEnvPath}
\alias{getEzZarrEnvPath}
\title{Get path to ez-zarr conda environment}
\usage{
getEzZarrEnvPath()
}
\value{
A character scalar indicating the path to the conda environment
provided by the package.
}
\description{
Get path to ez-zarr conda environment
}
\examples{
getEzZarrEnvPath()

}
\author{
Charlotte Soneson
}
