% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEzZarrVersion.R
\name{getEzZarrVersion}
\alias{getEzZarrVersion}
\title{Get names of ez-zarr functions}
\usage{
getEzZarrVersion()
}
\value{
A character scalar indicating the version of ez-zarr installed in
the package environment.
}
\description{
Get names of ez-zarr functions
}
\examples{
getEzZarrVersion()

}
\author{
Charlotte Soneson
}
