% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enableEzZarr.R
\name{enableEzZarr}
\alias{enableEzZarr}
\title{Enable ez-zarr}
\usage{
enableEzZarr()
}
\value{
A list with imported modules.
}
\description{
Enable ez-zarr
}
\examples{
enableEzZarr()

}
\author{
Charlotte Soneson
}
