% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEzZarrVersion.R
\name{getEzZarrVersion}
\alias{getEzZarrVersion}
\title{Get the version of ez-zarr python package}
\usage{
getEzZarrVersion()
}
\value{
A character scalar indicating the version of ez-zarr installed in
the conda environment provided with the package.
}
\description{
Get the version of ez-zarr python package
}
\examples{
getEzZarrVersion()

}
\author{
Charlotte Soneson
}
